requirements.txt
# API Framework
fastapi==0.115.0
uvicorn[standard]==0.34.0
pydantic==2.11.0
pydantic-settings==2.9.1

# Embedding Models
sentence-transformers==4.1.0
torch==2.7.0
transformers==4.52.0

# Utilities
python-dotenv==1.0.1
numpy==2.2.0
tqdm==4.66.0

# Testing
pytest==8.3.0
httpx==0.27.0
pytest-cov==4.1.0

# Performance
ujson==5.9.0
orjson==3.10.0
